Edit packaging
Packaging is important in keeping our code clean.
Move brandon.chatbot to src/main/java directory.
