+++ build.gradle
@@ -31,7 +31,7 @@ test {
 }
 
 application {
-    mainClass.set("Duke")
+    mainClass.set("brandon.chatbot.Duke")
 }
 
 shadowJar {
