--- src/brandon/chatbot/commands/UnknownCommand.java
@@ -1,8 +1,0 @@
-package commands;
-
-public class UnknownCommand extends Command {
-    @Override
-    public CommandResult execute() {
-        return new CommandResult("I am not intelligent enough to understand what that means...");
-    }
-}
