--- src/brandon/chatbot/common/DukeException.java
@@ -1,6 +1,0 @@
-package common;
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
