--- src/brandon/chatbot/common/DukeIndexOutOfBoundsException.java
@@ -1,7 +1,0 @@
-package common;
-
-public class DukeIndexOutOfBoundsException extends DukeException {
-    public DukeIndexOutOfBoundsException() {
-        super("    The index is way too big... I think...\n--------------------------------");
-    }
-}
