--- src/brandon/chatbot/common/DukeUnknownCommandException.java
@@ -1,8 +1,0 @@
-package common;
-public class DukeUnknownCommandException extends DukeException {
-
-    public DukeUnknownCommandException() {
-        super("    I am not intelligent enough to understand what that means...\n--------------------------------");
-    }
-}
-
