--- src/brandon/chatbot/tasks/Task.java
@@ -1,31 +1,0 @@
-package tasks;
-
-import common.DukeException;
-
-public class Task {
-    private String title;
-    private boolean done;
-
-    public Task(String title) throws DukeException {
-        if (title.isBlank()) {
-            throw new DukeException("    Title of a " + this + " cannot be blank...\n--------------------------------");
-        }
-        this.title = title;
-        this.done = false;
-    }
-
-    public boolean getDone() {
-        return this.done;
-    }
-    public void setDone(boolean done) {
-        this.done = done;
-    }
-
-    public String getStatus() {
-        return "[" + (this.done ? "X" : " ") + "] " + this.title;
-    }
-
-    public String toString() {
-        return title;
-    }
-}
