--- src/brandon/chatbot/tasks/Todo.java
@@ -1,21 +1,0 @@
-package tasks;
-
-import common.DukeException;
-
-public class Todo extends Task {
-
-    public Todo(String title) throws DukeException {
-        super(title);
-    }
-    @Override
-    public String toString() {
-        return "Todo";
-    }
-
-    @Override
-    public String getStatus() {
-        return "[T]" + super.getStatus();
-    }
-
-}
-
