+++ src/main/java/brandon/chatbot/commands/ExitCommand.java
@@ -1,0 +1,13 @@
+package brandon.chatbot.commands;
+
+public class ExitCommand extends Command {
+    public static final String COMMAND_WORD = "exit";
+    @Override
+    public CommandResult execute() {
+        return new CommandResult("bye bye...");
+    }
+
+    public static boolean isExit(Command command) {
+        return command instanceof ExitCommand; // instanceof returns false if it is null
+    }
+}
