+++ src/main/java/brandon/chatbot/commands/ListCommand.java
@@ -1,0 +1,11 @@
+package brandon.chatbot.commands;
+
+public class ListCommand extends Command {
+    public static final String LIST_SUCCESS = "ok... I'm listing..";
+
+    @Override
+    public CommandResult execute() {
+
+        return new CommandResult(LIST_SUCCESS, tasks);
+    }
+}
