+++ src/main/java/brandon/chatbot/commands/UnknownCommand.java
@@ -1,0 +1,8 @@
+package brandon.chatbot.commands;
+
+public class UnknownCommand extends Command {
+    @Override
+    public CommandResult execute() {
+        return new CommandResult("I am not intelligent enough to understand what that means...");
+    }
+}
