+++ src/main/java/brandon/chatbot/common/DukeException.java
@@ -1,0 +1,6 @@
+package brandon.chatbot.common;
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
