+++ src/main/java/brandon/chatbot/common/DukeIndexOutOfBoundsException.java
@@ -1,0 +1,7 @@
+package brandon.chatbot.common;
+
+public class DukeIndexOutOfBoundsException extends DukeException {
+    public DukeIndexOutOfBoundsException() {
+        super("    The index is way too big... I think...\n--------------------------------");
+    }
+}
