+++ src/main/java/brandon/chatbot/common/DukeUnknownCommandException.java
@@ -1,0 +1,8 @@
+package brandon.chatbot.common;
+public class DukeUnknownCommandException extends DukeException {
+
+    public DukeUnknownCommandException() {
+        super("    I am not intelligent enough to understand what that means...\n--------------------------------");
+    }
+}
+
